+++ projects/dt/Cargo.toml
@@ -13,3 +13,8 @@ clap = "3.0.0-beta.2"
 log = "0.4"
 env_logger = "0.8"
 tempfile = "3.1"
+git2 = "0.13.17"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+libc = "0.2"
+ansi_term = "0.12"
