+++ projects/dt/src/rprompt
@@ -1,0 +1,1 @@
+../../rprompt/src/rprompt
