+++ projects/rclio/Cargo.toml
@@ -7,8 +7,6 @@ publish = false
 
 [dependencies]
 ansi_term = "0.12"
-rprompt = { path = "../rprompt", version = "1.0" }
-rpassword =  { path = "../rpassword", version = "5.0" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
