+++ projects/rclio/src/lib.rs
@@ -1,6 +1,10 @@
 #[allow(unused)]
 mod rclio;
 #[allow(unused)]
+mod rpassword;
+#[allow(unused)]
+mod rprompt;
+#[allow(unused)]
 mod rutil;
 
 pub use crate::rclio::*;
