+++ projects/rclio/src/rclio/all.rs
@@ -1,11 +1,13 @@
+use crate::rpassword::{
+    prompt_password, prompt_password_from_bufread, read_password, read_password_from_bufread,
+};
+use crate::rprompt::{
+    prompt_reply, prompt_reply_from_bufread, read_reply, read_reply_from_bufread,
+};
 use crate::rutil::atty;
 use crate::rutil::safe_string::SafeString;
 use ansi_term::Color::{Green, Red, Yellow};
 use ansi_term::Style as AnsiTermStyle;
-use rpassword::{
-    prompt_password, prompt_password_from_bufread, read_password, read_password_from_bufread,
-};
-use rprompt::{prompt_reply, prompt_reply_from_bufread, read_reply, read_reply_from_bufread};
 use std::io::Result as IoResult;
 use std::io::{Cursor, StderrLock, StdinLock, StdoutLock, Write};
