+++ projects/rooster/Cargo.toml
@@ -19,8 +19,6 @@ rand = "0.7"
 libc = "0.2"
 byteorder = "1.3"
 clipboard = "0.5"
-rpassword =  { path = "../rpassword", version = "5.0" }
-rprompt = { path = "../rprompt", version = "1.0" }
 dirs = "3.0"
 openssl = "0.10"
 ansi_term = "0.12"
