+++ projects/rooster/src/rpassword
@@ -1,0 +1,1 @@
+../../rpassword/src/rpassword
