+++ projects/rooster/src/rpassword.rs
@@ -1,0 +1,1 @@
+../../rpassword/src/rpassword.rs
