+++ projects/rooster/src/rprompt.rs
@@ -1,0 +1,1 @@
+../../rprompt/src/rprompt.rs
