+++ projects/rpassword/src/rpassword.rs
@@ -1,0 +1,3 @@
+mod all;
+
+pub use all::*;
